LayoutTests/js/Promise-simple-in-workers-expected.txt

[Worker] Test Promise.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Starting worker: resources/Promise-simple-in-workers.js
PASS [Worker] thisInInit === firstPromise is true
PASS [Worker] thisInFulfillCallback === secondPromise is true
PASS [Worker] result is "hello"
PASS successfullyParsed is true

TEST COMPLETE
